+++ src/lib/pages/account-details/components/tables/AdminContractsTable.tsx
@@ -1,12 +1,13 @@
 import { Box } from "@chakra-ui/react";
 import type { ChangeEvent } from "react";
 
+import { useInternalNavigate } from "lib/app-provider";
 import { Pagination } from "lib/components/pagination";
 import { usePaginator } from "lib/components/pagination/usePaginator";
 import { EmptyState } from "lib/components/state";
 import { ContractsTable, TableTitle, ViewMore } from "lib/components/table";
 import { useAccountAdminContracts } from "lib/pages/account-details/data";
-import type { HumanAddr, Option } from "lib/types";
+import type { ContractAddr, HumanAddr, Option } from "lib/types";
 
 interface AdminContractsTableProps {
   walletAddress: HumanAddr;
@@ -23,6 +24,13 @@ export const AdminContractsTable = ({
   refetchCount,
   onViewMore,
 }: AdminContractsTableProps) => {
+  const navigate = useInternalNavigate();
+  const onRowSelect = (contract: ContractAddr) =>
+    navigate({
+      pathname: "/contract/[contract]",
+      query: { contract },
+    });
+
   const {
     pagesQuantity,
     currentPage,
@@ -68,10 +76,12 @@ export const AdminContractsTable = ({
         isLoading={isLoading}
         emptyState={
           <EmptyState
+            imageVariant="empty"
             message="This account does not have any admin access for any contracts."
             withBorder
           />
         }
+        onRowSelect={onRowSelect}
       />
       {!!totalData &&
         (onViewMore